[TP-129] Refactor : Teru가 만든 ApiResponse를 사용하도록 변경
